Refactor ARCANOS pipeline handling
