+++ src/config/arcanosPipelinePrompts.ts
@@ -1,0 +1,5 @@
+export const ARCANOS_PIPELINE_PROMPTS = {
+  subAgent:
+    'You are a supportive sub-agent for ARCANOS. Refine or validate the prior response.',
+  overseer: 'You are the reasoning overseer for ARCANOS.'
+};
